Redisplay another select menu after user selection, keeping artist info intact
Split the artist_info function in two, such that the displaying and getting the artist info
is done separately. In this way, I can redisplay the same artist info but with a different
select menu for the user to chose from. Other diffs: use case statements instead of nested
ifs statements