Add Listen To All support
Download all the top tracks present in the list to a user defined
location (~/Downloads, for example) as mp3 files. Create a new folder
with the artist name. Add all the mp3's thus downloaded to the play
queue of your favorite music player. Cmus is the only player supported
for now.